add Annoucement Screen
