+++ sync-trip/App.tsx
@@ -1,30 +1,39 @@
 import React from 'react';
-import {Provider as PaperProvider} from 'react-native-paper';
+import { Provider as PaperProvider } from 'react-native-paper';
 
-import {MessageDialogProvider} from './components/MessageDialog';
+import { MessageDialogProvider } from './components/MessageDialog';
 import AppNavigator from './navigation/AppNavigator';
 
-import {en, registerTranslation} from 'react-native-paper-dates';
+import { en, registerTranslation } from 'react-native-paper-dates';
 import 'expo-dev-client';
 import './utils/firebase';
 import 'react-native-get-random-values';
 
 import NotificationHandler from './utils/NotificationHandler';
 
-import {NotificationProvider} from './context/NotificationContext';
+import { NotificationProvider } from './context/NotificationContext';
+import { UserProvider } from "./context/UserContext";
+import { TripProvider } from "./context/TripContext";
+import { BillTransactionProvider } from "./context/BillAndTransactionContext";
 
 registerTranslation('en', en);
 
 const App = () => {
   return (
-    <PaperProvider>
-      <MessageDialogProvider>
-        <NotificationProvider>
-          <NotificationHandler/>
-          <AppNavigator/>
-        </NotificationProvider>
-      </MessageDialogProvider>
-    </PaperProvider>
+    <UserProvider>
+      <TripProvider>
+        <BillTransactionProvider>
+          <PaperProvider>
+            <MessageDialogProvider>
+              <NotificationProvider>
+                <NotificationHandler />
+                <AppNavigator />
+              </NotificationProvider>
+            </MessageDialogProvider>
+          </PaperProvider>
+        </BillTransactionProvider>
+      </TripProvider>
+    </UserProvider>
   );
 };
